requirements.txt
asgiref==3.8.1
Django==5.0.7
djangorestframework==3.15.2
sqlparse==0.5.1
typing_extensions==4.12.2
django-ckeditor==6.7.1
pillow==10.4.0
django-colorfield==0.11.0
django-cors-headers==4.3.1
drf-spectacular==0.27.1
requests==2.31.0